% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreport.R
\name{calendar_autoreport}
\alias{calendar_autoreport}
\title{Make calendar of report schedule and daily load}
\usage{
calendar_autoreport(data, pointRangeMax = 0)
}
\arguments{
\item{data}{List containing auto report data
run. May well contain repeating values}

\item{pointRangeMax}{Integer to provide static range of geom_point as
\code{range(1:pointRangeMax)}. If set to 0 (default) the range of current data will be
used.}
}
\value{
a (gg)plot object
}
\description{
Make calendar of report schedule and daily load
}
