% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoreport.R
\name{unique_autoreport}
\alias{unique_autoreport}
\title{Tools for handling auto report data}
\usage{
unique_autoreport(data, entety)
}
\arguments{
\item{data}{List of auto report data}

\item{entety}{A character string defining the entety. Likely one of
\code{c("package", "type", "owner", "organization")}.}
}
\value{
A character vector of unique values for the entity in question.
}
\description{
Tools for handling auto report data
}
\examples{
ar <- list(ar1 = list(type = "A"), ar2 = list(type = "B"))
unique_autoreport(ar, "type") # c("A", "B")
}
